+++ app-backend/api/src/main/scala/datasource/QueryParameters.scala
@@ -10,6 +10,6 @@ import com.azavea.rf.api.utils.queryparams._
 
 trait DatasourceQueryParameterDirective extends QueryParametersCommon {
   def datasourceQueryParams = (
-    searchQueryParams
+    searchParams
   ).as(DatasourceQueryParameters.apply _)
 }
