+++ app-backend/api/src/main/scala/tools/Routes.scala
@@ -6,7 +6,6 @@ import com.azavea.rf.datamodel._
 import com.azavea.rf.tool.ast._
 import com.azavea.rf.tool.ast.codec._
 import com.azavea.rf.database.filter.Filterables._
-import com.azavea.rf.api.utils.queryparams.QueryParametersCommon
 import com.azavea.maml.serve._
 import io.circe._
 import akka.http.scaladsl.model.StatusCodes
@@ -28,7 +27,7 @@ import doobie.postgres.implicits._
 
 
 trait ToolRoutes extends Authentication
-    with QueryParametersCommon
+    with ToolQueryParameterDirective
     with PaginationDirectives
     with CommonHandlers
     with KamonTraceDirectives
