+++ app-backend/api/src/main/scala/utils/QueryParameters.scala
@@ -30,7 +30,7 @@ trait QueryParametersCommon extends QueryParameterDeserializers {
     orgQueryParams &
     userQueryParameters &
     timestampQueryParameters &
-    searchQueryParams
+    searchParams
   ).as(ProjectQueryParameters.apply _)
 
   def aoiQueryParameters = (
@@ -73,23 +73,7 @@ trait QueryParametersCommon extends QueryParameterDeserializers {
     orgQueryParams & userQueryParameters & timestampQueryParameters
   ).as(ShapeQueryParameters.apply _)
 
-  def searchQueryParams = parameters(
-    'searchPattern.as[String].?
-  ).as(SearchQueryParameters.apply _)
-
-  def toolParameters = parameters(
-    'minRating.as[Double].?,
-    'maxRating.as[Double].?,
-    'toolCategory.as[String].*,
-    'toolTag.as[String].*,
+  def searchParams = parameters(
     'search.as[String].?
-  ).as(ToolQueryParameters.apply _)
-
-  def combinedToolQueryParams = (
-    orgQueryParams &
-    userQueryParameters &
-    timestampQueryParameters &
-    toolParameters &
-    searchQueryParams
-  ).as(CombinedToolQueryParameters.apply _)
+  ).as(SearchQueryParameters.apply _)
 }
