+++ app-backend/db/src/main/scala/com/azavea/rf/database/filters/Filters.scala
@@ -65,6 +65,4 @@ object Filters {
       })
     )
   }
-
-
 }
